requirements.txt
discord.py
python-dotenv
requests
time
PyNaCl